Extract TraktItem to own file
